Moved components specs to "components-specs" folder
